apps/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
scipy==1.11.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pymongo==4.6.0
elasticsearch==8.11.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2